pyproject.toml

[tool.poetry]
name = "ekuru_ltn"

packages = [
    { include = "ltn_imp" },
]

version = "2.16.0"
description = "A parser for FOL that utilizes NLTK Logic Module to parse and fuzzify FOL rules."
authors = ["Emre Kuru <emre.kuru@ozu.edu.tr>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.11.0, < 4.0.0"
torch = "^2.3.1"
nltk = "^3.8.1"
nbconvert = "^7.16.4"
nbformat = "^5.10.4"
pyparsing = "^3.1.2"
tensorboard = "^2.17.0"
torchviz = "^0.0.2"
pyyaml = "^6.0.2"
tatsu = "^5.12.1"

[tool.poetry.group.dev.dependencies]
poetry = "^1.7.0"
pytest = "^8.1.0"
coverage = "^7.4.0"
mypy = "^1.9.0"
gitpython = "^3.1.43"
ipykernel = "^6.29.5"
matplotlib = "^3.9.0"
ltntorch = "^1.0.1"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
poethepoet = "^0.27.0"

[tool.poe.tasks]
create-dataset-dir = { shell = "mkdir -p examples/datasets" }
download-iris-training = { shell = "curl -L -o examples/datasets/iris_training.csv https://raw.githubusercontent.com/tommasocarraro/LTNtorch/main/examples/datasets/iris_training.csv" }
download-iris-test = { shell = "curl -L -o examples/datasets/iris_test.csv https://raw.githubusercontent.com/tommasocarraro/LTNtorch/main/examples/datasets/iris_test.csv" }
download-real_estate = { shell = "curl -L -o examples/datasets/real_estate.csv https://raw.githubusercontent.com/tommasocarraro/LTNtorch/main/examples/datasets/real-estate.csv" }
download-datasets = { sequence = ["create-dataset-dir", "download-iris-training", "download-iris-test", "download-real_estate"] }

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true